Pagination search + auth error handler + spinner + block button + sync playlist

+++ api/api.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const log = require('../log/log.js');
 const youtubeAPI = require('./youtube-api.js');
 const vimeoAPI = require('./vimeo-api.js');
 const got = require('got');
+const path = require('path');
 
 const userApiUrl = "http://localhost:3001/";
 const playlistApiUrl ="http://localhost:3003/";
@@ -13,51 +13,66 @@ router.get('/user/:token', function(req, res) {
     got('/user/'+req.params.token, { 
         baseUrl: userApiUrl, 
         json: true })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
     
 });
 
 router.get('/users/:id', function(req, res) {
-    got('/users/'+req.params.id, { 
+    got('user/all/'+req.params.id, { 
         baseUrl: userApiUrl, 
         json: true })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 router.post('/authenticate', function(req, res){
-    got('/authenticate', { 
+    got('/user/authenticate', { 
         baseUrl: userApiUrl, 
         json: true,
         body: req.body })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
+    .then(response => {
+        res.send(response.body)}
+        )
+    .catch((error) => {
+        console.log(error);
+        res.status(error.statusCode).send(error)
     });
 });
 
 router.post('/register', function(req,res) {
-    got('/register', { 
+    got('/user/register', { 
         baseUrl: userApiUrl, 
         json: true,
         body: req.body })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
+});
+
+router.post('/user/admin', function(req,res) {
+    got('/user/admin', { 
+        baseUrl: userApiUrl, 
+        json: true,
+        body: req.body })
+    .then(response => res.send(response.body))
+    .catch(handleError);
+});
+
+router.get('/user/blocked/:id', function(req,res) {
+    got('/user/blocked/'+req.params.id, { 
+        baseUrl: userApiUrl, 
+        json: true })
+    .then(response => res.send(response.body))
+    .catch(handleError);
+});
+
+router.post('/user/blocked', function(req,res) {
+    got('/user/blocked', { 
+        baseUrl: userApiUrl, 
+        json: true,
+        body: req.body })
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 
@@ -67,24 +82,16 @@ router.post('/playlist', function(req,res) {
         baseUrl: playlistApiUrl, 
         json: true,
         body: req.body })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 router.get('/playlist/:id', function(req, res) {  
     got('/playlist/'+req.params.id, { 
         baseUrl: playlistApiUrl, 
         json: true })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 router.post('/playlist/video', function(req, res) {
@@ -92,12 +99,8 @@ router.post('/playlist/video', function(req, res) {
         baseUrl: playlistApiUrl, 
         json: true,
         body: req.body })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 router.delete('/playlist/:id', function(req, res) {
@@ -105,59 +108,74 @@ router.delete('/playlist/:id', function(req, res) {
         baseUrl: playlistApiUrl, 
         json: true,
         method: 'DELETE' })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 router.post("/search", function(req, res) {
     if(req.body.api == "YouTube"){
         youtubeAPI.search(req,res);
     }
-
     if(req.body.api == "Vimeo"){
         vimeoAPI.searchListByKeyword(req, res);
     }
 });
 
+router.post("/page", function(req, res) {
+    if(req.body.api == "YouTube"){
+        youtubeAPI.page(req,res);
+    }
+    if(req.body.api == "Vimeo"){
+        vimeoAPI.page(req, res);
+    }
+});
+
 router.get("/log/search/:id", function(req, res) {
-    got('/log/search'+req.params.id, { 
+    got('/log/search/'+req.params.id, { 
         baseUrl: logApiUrl, 
         json: true })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
+    
 });
 
 router.get("/log/:id", function(req, res) {
-    got('/log/'+req.params.id, { 
-        baseUrl: logApiUrl, 
+    got('/user/admin/'+req.params.id, {
+        baseUrl: userApiUrl,
         json: true })
-    .then(function(response) { 
-        res.send(response.body);
+    .then(function(response) {
+        if(response.body.isAdmin){
+            got('/log/'+req.params.id, { 
+                baseUrl: logApiUrl, 
+                json: true })
+            .then(response => res.send(response.body))
+            .catch(handleError);
+        }
+        else {
+            res.send();
+        }
     })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .catch(handleError);
 });
 
 router.delete('/log/search/:id', function(req, res) {
-    got('/log/search/'+req.params.id, { 
-        baseUrl: logApiUrl, 
-        json: true,
-        method: 'DELETE' })
-    .then(function(response) { 
-        res.send(response.body);
+    got('/user/admin/'+req.params.id, {
+        baseUrl: userApiUrl,
+        json: true })
+    .then(function(response) {
+        if(response.body.isAdmin){
+            got('/log/search/'+req.params.id, { 
+                baseUrl: logApiUrl, 
+                json: true,
+                method: 'DELETE' })
+            .then(response => res.send(response.body))
+            .catch(handleError);
+        }
+        else {
+            res.send();
+        }
     })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .catch(handleError);
 });
 
 router.delete('/log/:id', function(req, res) {
@@ -165,25 +183,44 @@ router.delete('/log/:id', function(req, res) {
         baseUrl: logApiUrl, 
         json: true,
         method: 'DELETE' })
-    .then(function(response) { 
-        res.send(response.body);
-    })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .then(response => res.send(response.body))
+    .catch(handleError);
 });
 
 router.get('/logout/:id', function(req, res){
     got('/logout/'+req.params.id, { 
         baseUrl: logApiUrl, 
         json: true })
-    .then(function(response) { 
+    .then(response => res.send(response.body))
+    .catch(handleError);
+});
+
+router.post('/forgot', function (req, res) {    
+    got('/forgot', {
+        baseUrl: userApiUrl,
+        json: true,
+        body: req.body
+    }).then(function(response) { 
         res.send(response.body);
     })
-    .catch(function(error){
-        console.log('error:', error);
-    });
+    .catch(handleError);
 });
+
+router.post('/resetpw/:token',function(req, res) {
+    console.log("post reset : "+req.params.token);
+    got('/reset/'+req.params.token, {
+        baseUrl: userApiUrl,
+        json: true,
+        body: req.body
+    }).then(function(response) { 
+        res.send(response.body);
+    })
+    .catch(handleError);
+});
+
+function handleError(error){
+    console.log('error:', error);
+}
 
 module.exports = {
     router: router,
